Addressing some of Charlie's suggestions
