+++ lib/util/browserdetection.js
@@ -18,6 +18,17 @@ function isIOS() {
 }
 
 /**
+ * Detects whether or not a device is an iPad
+ * @returns {boolean}
+ */
+
+function isIpad() {
+  return navigator && navigator.maxTouchPoints &&
+  navigator.maxTouchPoints > 2 &&
+  (/Macintosh/i.test(navigator.userAgent) || /iPad/.test(navigator.userAgent) || /iPad/.test(navigator.platform));
+}
+
+/**
  * Check whether the current browser is a mobile browser
  * @returns {boolean}
  */
@@ -123,6 +134,7 @@ function getParenthesizedSubstrings(string) {
 module.exports = {
   isAndroid,
   isIOS,
+  isIpad,
   isMobile,
   isNonChromiumEdge,
   mobileWebKitBrowser,
