+++ lib/util/index.js
@@ -134,17 +134,6 @@ function flatMap(list, mapFn) {
 }
 
 /**
- * Detects whether or not a device is an iPad
- * @returns {boolean}
- */
-
-function isIpad() {
-  return navigator && navigator.maxTouchPoints &&
-  navigator.maxTouchPoints > 2 &&
-  (/Macintosh/i.test(navigator.userAgent) || /iPad/.test(navigator.userAgent) || /iPad/.test(navigator.platform));
-}
-
-/**
  * Get the user agent string, or return "Unknown".
  * @returns {string}
  */
@@ -873,7 +862,6 @@ exports.difference = difference;
 exports.filterObject = filterObject;
 exports.flatMap = flatMap;
 exports.getPlatform = getPlatform;
-exports.isIpad = isIpad;
 exports.getUserAgent = getUserAgent;
 exports.hidePrivateProperties = hidePrivateProperties;
 exports.hidePrivateAndCertainPublicPropertiesInClass = hidePrivateAndCertainPublicPropertiesInClass;
