+++ lib/util/insightspublisher/index.js
@@ -13,6 +13,7 @@ const toplevel = globalThis;
 const WebSocket = toplevel.WebSocket ? toplevel.WebSocket : require('ws');
 const { PUBLISHER_METADATA_IPAD } = require('../constants');
 const util = require('../../util');
+const { isIpad } = require('../browserdetection');
 
 /**
  * Publish events to the Insights gateway.
@@ -62,9 +63,6 @@ class InsightsPublisher extends EventEmitter {
       },
       _WebSocket: {
         value: options.WebSocket
-      },
-      _isIpad: {
-        value: util.isIpad()
       }
     });
 
@@ -202,7 +200,7 @@ function connect(publisher, token, sdkName, sdkVersion, roomSid, participantSid,
       roomSid: roomSid,
     };
 
-    if (publisher._isIpad) {
+    if (isIpad()) {
       connectRequest.publisher = { ...connectRequest.publisher, ...PUBLISHER_METADATA_IPAD };
       // connectRequest.publisher = Object.assign(connectRequest.publisher, PUBLISHER_METADATA_IPAD);
     }
